backend/requirements.txt
Flask==3.0.0
numpy==1.26.0
Pillow==10.0.1
tensorflow==2.15.0 # For CNN Model
keras==2.15.0
Flask-Cors==4.0.0
